requirements.txt
squareup
python-dotenv
pandas==2.2.2
#dataclasses
#logging